(#11095) [icu] Eliminate the runConfigureICU script from the ICU build.
* Eliminate the runConfigureICU script from the ICU build.

runConfigureICU sets up a few fixed configuraitons, providing flags to configure, but also
prevents configure from using other environment variables it is designed to use.
We could, if there is a need, add these to flags set in conanfile.py, but because
Conan supports Autoconf directly, there probably isn't a need for this.
The reciple only supports a few configuraitons, so adding support for a new compiler
or operating system would require modificaitons to runConfgiureICU - none of which is actually useful.

* Mimic one thing that runConfigureICU did, by setting CC and CXX to cl when msvc or visual studio is the compiler.

* Add -FS to Visual Studio/MSVC builds.